+++ gd-gltg-bmp/webpack.config.prod.js
@@ -12,7 +12,8 @@ module.exports = webpackMerge.merge(
         plugins: [
             new Webpack.DefinePlugin({
                 'process.env.NODE_ENV': '"production"',
-                'process.env.GEOSERVER_URL': JSON.stringify(process.env.GEOSERVER_URL || '/geoserver')
+                'process.env.GEOSERVER_URL': JSON.stringify(process.env.GEOSERVER_URL || '/geoserver'),
+                'process.env.BMP_API_URL': JSON.stringify(process.env.BMP_API_URL || '/bmp')
             })
         ]
     }